requirements.txt
fastapi
uvicorn
requests
numpy
pydantic
python-dotenv
